# openapi/api/openapi.yaml
openapi: 3.0.3
info:
  contact:
    name: arimitsu-koichiro
  description: sample
  title: sample Backend
  version: 1.0.0
servers:
  - url: /
paths:
...
  /api/v1/channel/{channel_id}:
    get:
      operationId: subscribe_channel
      parameters:
        - explode: false
          in: path
          name: channel_id
          required: true
          schema:
            type: string
          style: simple
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelMessage'
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
          description: デフォルトのエラーレスポンス
      tags:
        - channel
    post:
      operationId: publish_channel
      parameters:
        - explode: false
          in: path
          name: channel_id
          required: true
          schema:
            type: string
          style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChannelMessage'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusOk'
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
          description: デフォルトのエラーレスポンス
      tags:
        - channel
...
components:
...
  schemas:
    StatusOk:
      example:
        status: OK
      properties:
        status:
          default: OK
          type: string
      required:
        - status
      type: object
...
    ErrorMessage:
      properties:
        status:
          type: integer
        type:
          type: string
        message:
          type: string
      required:
        - status
        - type
      type: object
...
    ChannelMessage:
      example:
        payload: payload
        topic: topic
      properties:
        topic:
          type: string
        payload:
          type: string
      required:
        - payload
        - topic
      type: object